backend/requirements.txt
Django>=5.1.4,<5.2
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.10,<2.10
drf-spectacular>=0.28.0,<0.29
Pillow>=11.1.0,<11.2
django-jet-reboot>=1.3.10,<1.4
djangorestframework-simplejwt>=5.3.1,<5.4
django-phonenumber-field>=8.0.0,<8.1